requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
hiredis==2.0.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
ldap3==2.9.1
Markdown==3.3.7
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1
zipp==3.8.1
